requirements.txt
PyGithub
pydantic>=2.0.0
pydantic-settings
Jinja2